Use SkinMustache and merge CitizenTemplate into SkinCitizen
